psp-3415 ltsa and property calls now using api wrapper. updated sidebar loading logic to use api wrapper boolean. display an error when ltsa load completes with no data. ensure that correct tab is displayed when no property information is available.
