+++ frontend/src/features/mapSideBar/MotiInventoryHeader.tsx
@@ -12,6 +12,8 @@ import { HeaderField } from './tabs/HeaderField';
 
 interface IMotiInventoryHeaderProps {
   ltsaData?: LtsaOrders;
+  ltsaLoading: boolean;
+  propertyLoading: boolean;
   property?: IPropertyApiModel;
   onZoom?: (apiProperty?: IPropertyApiModel | undefined) => void;
 }
@@ -24,7 +26,7 @@ export const MotiInventoryHeader: React.FunctionComponent<IMotiInventoryHeaderPr
       : (props.ltsaData?.parcelInfo.orderedProduct.fieldedData
           .associatedPlans as AssociatedPlan[]).map(x => x.planNumber);
 
-  const isLoading = props.ltsaData === undefined || props.property === undefined;
+  const isLoading = props.ltsaLoading || props.propertyLoading;
   return (
     <>
       <LoadingBackdrop show={isLoading} parentScreen={true} />
