+++ frontend/src/features/mapSideBar/tabs/ltsa/LtsaChargeOwnerSubForm.test.tsx
@@ -1,10 +1,10 @@
 import { Formik } from 'formik';
 import { LtsaOrders } from 'interfaces/ltsaModels';
 import { noop } from 'lodash';
+import { mockLtsaResponse } from 'mocks/filterDataMock';
 import { render, RenderOptions } from 'utils/test-utils';
 
 import LtsaChargeOwnerSubForm, { ILtsaChargeOwnerSubFormProps } from './LtsaChargeOwnerSubForm';
-import { mockLtsaResponse } from './LtsaTabView.test';
 
 describe('LtsaChargeOwnerSubForm component', () => {
   const setup = (
