+++ frontend/src/features/mapSideBar/tabs/ltsa/LtsaChargeSubForm.test.tsx
@@ -1,10 +1,10 @@
 import { Formik } from 'formik';
 import { LtsaOrders } from 'interfaces/ltsaModels';
 import { noop } from 'lodash';
+import { mockLtsaResponse } from 'mocks/filterDataMock';
 import { render, RenderOptions } from 'utils/test-utils';
 
 import LtsaChargeSubForm, { ILtsaChargeSubFormProps } from './LtsaChargeSubForm';
-import { mockLtsaResponse } from './LtsaTabView.test';
 
 describe('LtsaChargeSubForm component', () => {
   const setup = (
