+++ frontend/src/features/mapSideBar/tabs/ltsa/LtsaLandSubForm.test.tsx
@@ -1,10 +1,10 @@
 import { Formik } from 'formik';
 import { LtsaOrders } from 'interfaces/ltsaModels';
 import { noop } from 'lodash';
+import { mockLtsaResponse } from 'mocks/filterDataMock';
 import { render, RenderOptions } from 'utils/test-utils';
 
 import LtsaLandSubForm, { ILtsaLandSubFormProps } from './LtsaLandSubForm';
-import { mockLtsaResponse } from './LtsaTabView.test';
 
 describe('LtsaLandSubForm component', () => {
   const setup = (
