+++ frontend/src/features/mapSideBar/tabs/ltsa/LtsaOwnershipInformationForm.test.tsx
@@ -1,12 +1,12 @@
 import { Formik } from 'formik';
 import { LtsaOrders } from 'interfaces/ltsaModels';
 import { noop } from 'lodash';
+import { mockLtsaResponse } from 'mocks/filterDataMock';
 import { render, RenderOptions } from 'utils/test-utils';
 
 import LtsaOwnershipInformationForm, {
   ILtsaOwnershipInformationFormProps,
 } from './LtsaOwnershipInformationForm';
-import { mockLtsaResponse } from './LtsaTabView.test';
 
 describe('LtsaOwnershipInformationForm component', () => {
   const setup = (
