+++ frontend/src/features/mapSideBar/tabs/ltsa/LtsaTransferSubForm.test.tsx
@@ -1,9 +1,9 @@
 import { Formik } from 'formik';
 import { LtsaOrders } from 'interfaces/ltsaModels';
 import { noop } from 'lodash';
+import { mockLtsaResponse } from 'mocks/filterDataMock';
 import { render, RenderOptions } from 'utils/test-utils';
 
-import { mockLtsaResponse } from './LtsaTabView.test';
 import LtsaTransferSubForm, { ILtsaTransferSubFormProps } from './LtsaTransferSubForm';
 
 describe('LtsaTransferSubForm component', () => {
