+++ frontend/src/hooks/useLtsa.ts
@@ -1,30 +1,29 @@
+import { AxiosResponse } from 'axios';
+import { LtsaOrders } from 'interfaces/ltsaModels';
 import { useCallback } from 'react';
-import { useDispatch } from 'react-redux';
 import { toast } from 'react-toastify';
-import { handleAxiosResponse } from 'utils';
+import { pidFormatter } from 'utils';
 
 import { useApiLtsa } from './pims-api/useApiLtsa';
+import { useApiRequestWrapper } from './pims-api/useApiRequestWrapper';
 
 /**
  * hook retrieves data from ltsa
  */
 export const useLtsa = () => {
   const { getLtsaOrders } = useApiLtsa();
-  const dispatch = useDispatch();
 
-  const getLtsaData = useCallback(
-    async (pid: string) => {
-      try {
-        const response = await handleAxiosResponse(dispatch, 'getLtsaData', getLtsaOrders(pid));
-        return response;
-      } catch (axiosError) {
-        toast.error(
-          `Failed to get LTSA data. error from LTSA: ${axiosError?.response?.data.error}`,
-        );
-      }
-    },
-    [dispatch, getLtsaOrders],
-  );
+  const { execute, loading } = useApiRequestWrapper<
+    (pid: string) => Promise<AxiosResponse<LtsaOrders>>
+  >({
+    requestFunction: useCallback(async (pid: string) => await getLtsaOrders(pidFormatter(pid)), [
+      getLtsaOrders,
+    ]),
+    requestName: 'getLtsaData',
+    onError: useCallback(axiosError => {
+      toast.error(`Failed to get LTSA data. error from LTSA: ${axiosError?.response?.data.error}`);
+    }, []),
+  });
 
-  return { getLtsaData };
+  return { getLtsaData: execute, ltsaLoading: loading };
 };
